voters: Remove unreachable voters
Since we're switching to OpaVote, we need to take some care to not
add too many voters whose emails are bouncing, because the
[terms](https://opavote.com/terms) say

> 3.4 You must ensure that voter lists are limited to people who are
> genuinely interested in the subject matter of the election, that voter
> email addresses are usable (e.g., not bouncing), and that voters do not
> mark emails as spam. For any election, if more than 2% of emails are
> undeliverable (e.g., bouncing) or more than 0.1% of voters report an
> election email as spam, then at our sole discretion, (a) your election
> may be terminated immediately and without any refund of amounts paid, or
> (b) you may be required to pay a fee of US$1 for each bouncing email and
> US$5 for each email marked as spam by a voter.

and OpaVote doesn't have a separate email activation step like CIVS (and
we don't want to do that ourselves this time for now).

In addition, as we can redownload the voter list in OpaVote, we can also
afford to add people during the election and cross check at the end.